refactor(todo): update todo controller
- remove library specific approach for response returns
- add logic handles for service calls

+++ src/todo/todo.controller.ts
@@ -1,8 +1,8 @@
-import { Body, Controller, Delete, Get, HttpStatus, Param, Patch, Post, Res } from '@nestjs/common';
-import { Response } from 'express';
+import { Body, Controller, Delete, Get, InternalServerErrorException, NotFoundException, Param, Patch, Post, UnprocessableEntityException } from '@nestjs/common';
 import { plainToClass } from 'class-transformer';
 import { CreateTodoDTO, ModifyTodoDTO, Todo } from './todo.schema';
 import { TodoService } from './todo.service';
+import { ObjectId } from 'bson';
 
 @Controller('todos')
 export class TodoController {
@@ -10,66 +10,51 @@ export class TodoController {
   constructor (private readonly todoService: TodoService) {}
 
   @Post()
-  async create (@Body() createTodoDTO: CreateTodoDTO, @Res() res: Response) {
+  async create (@Body() createTodoDTO: CreateTodoDTO) {
     const createdTodo = await this.todoService.create(createTodoDTO);
-    return res
-      .status(HttpStatus.CREATED)
-      .json({
-        status: HttpStatus.CREATED,
-        items: plainToClass(Todo, createdTodo.toObject()),
-      })
-      .end();
+    if (!createdTodo) throw new InternalServerErrorException();
+
+    return plainToClass(Todo, createdTodo.toObject());
   }
 
   @Patch(':id')
-  async update (@Param('id') todoId: string, @Body() updateTodoDTO: ModifyTodoDTO, @Res() res: Response) {
+  async update (@Param('id') todoId: string, @Body() updateTodoDTO: ModifyTodoDTO) {
+    if (!todoId || !ObjectId.isValid(todoId)) { throw new UnprocessableEntityException(`${todoId} is not a valid id.`); }
+
     const updatedTodo = await this.todoService.update(todoId, updateTodoDTO);
-    return res
-      .status(HttpStatus.CREATED)
-      .json({
-        status: HttpStatus.CREATED,
-        items: plainToClass(Todo, updatedTodo.toObject()),
-      })
-      .end();
+    if (!updatedTodo) throw new NotFoundException();
+
+    return plainToClass(Todo, updatedTodo.toObject());
   }
 
   @Get()
-  async find (@Res() res: Response) {
-    const todos = await this.todoService.find();
+  async find () {
+    const todos = await this.todoService.find({});
     const serializedTodos = plainToClass(Todo, todos?.map(todo => todo.toObject()));
-    return res
-      .status(HttpStatus.OK)
-      .json({
-        status: HttpStatus.OK,
-        items: serializedTodos,
-      })
-      .end();
+
+    return serializedTodos;
   }
 
   @Get(':id')
-  async get (@Param('id') todoId: string, @Res() res: Response) {
+  async get (@Param('id') todoId: string) {
+    if (!todoId || !ObjectId.isValid(todoId)) { throw new UnprocessableEntityException(`${todoId} is not a valid id.`); }
+
     const todo = await this.todoService.findById(todoId);
+    if (!todo) throw new NotFoundException('Todo not found');
+
     const serializedTodo = plainToClass(Todo, todo?.toObject());
-    return res
-      .status(HttpStatus.OK)
-      .json({
-        status: HttpStatus.OK,
-        items: serializedTodo,
-      })
-      .end();
+    return serializedTodo;
   }
 
   @Delete(':id')
-  async remove (@Param('id') todoId: string, @Res() res: Response) {
+  async remove (@Param('id') todoId: string) {
+    if (!todoId || !ObjectId.isValid(todoId)) { throw new UnprocessableEntityException(`${todoId} is not a valid id.`); }
+
     const todo = await this.todoService.remove(todoId);
+    if (!todo) throw new NotFoundException();
+
     const serializedDeletedTodo = plainToClass(Todo, todo.toObject());
-    return res
-      .status(HttpStatus.OK)
-      .json({
-        status: HttpStatus.OK,
-        items: serializedDeletedTodo,
-      })
-      .end();
+    return serializedDeletedTodo;
   }
 
 }
